requirements.txt
certifi==2018.11.29
chardet==3.0.4
config==0.4.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.3.0
psycopg2==2.7.6.1
python-decouple==3.1
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
whitenoise==3.3.1
